v0.4.870

+++ gamestream_flutter/lib/isometric/edit_state.dart
@@ -21,7 +21,7 @@ final edit = EditState();
 class SpawnNodeData {
   final int spawnType;
   final int spawnAmount;
-  final double spawnRadius;
+  final int spawnRadius;
 
   SpawnNodeData({
     required this.spawnType,
